pyproject.toml
[tool.poetry]
name = "drf-ecommerce"
version = "0.1.0"
description = ""
authors = ["ZTZ <tzamirov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
coverage = "^7.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
